Update matcher to use in64 iids

+++ recordmatcherapi.go
@@ -9,10 +9,10 @@ import (
 
 // Match - force matches a record
 func (s *Server) Match(ctx context.Context, req *pb.MatchRequest) (*pb.MatchResponse, error) {
-	return &pb.MatchResponse{}, s.processRecordList(ctx, []int32{req.GetInstanceId()}, "api", req.GetForce())
+	return &pb.MatchResponse{}, s.processRecordList(ctx, []int64{req.GetInstanceId()}, "api", req.GetForce())
 }
 
 //ClientUpdate on an updated record
 func (s *Server) ClientUpdate(ctx context.Context, req *rcpb.ClientUpdateRequest) (*rcpb.ClientUpdateResponse, error) {
-	return &rcpb.ClientUpdateResponse{}, s.processRecordList(ctx, []int32{req.GetInstanceId()}, "capi", false)
+	return &rcpb.ClientUpdateResponse{}, s.processRecordList(ctx, []int64{req.GetInstanceId()}, "capi", false)
 }
